+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed an issue where some Xenoblade 3 shader database entries incorrectly assigned normal maps to glossiness.
 * Fixed an issue where some UV scale and material parameter values would be incorrect due to incorrectly applying callbacks.
 * Fixed an issue where some toon materials would render as black in the viewport due to incorrect specular rendering.
+* Fixed an issue where material parameters would not always include all values.
 
 ## 0.10.0 - 2024-06-18
 ### Added